requirements.txt
langchain
langgraph
pytest
chromadb
openai
mcp
langgraph-checkpoint-postgres  # For persistent learning
python-dotenv  # For loading .env files

